+++ index.php
@@ -15,6 +15,7 @@
 get_header();
 
 ?>
+	<?php echo onepress_breadcrumb(); ?>
 	<div id="content" class="site-content">
 		<div id="content-inside" class="container right-sidebar">
 			<div id="primary" class="content-area">
@@ -53,7 +54,7 @@ get_header();
 
 				</main><!-- #main -->
 			</div><!-- #primary -->
-			
+
 			<?php get_sidebar(); ?>
 
 		</div><!--#content-inside -->